backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
psutil==5.9.8
scikit-learn==1.4.2
numpy==1.26.4
mysql-connector-python==8.4.0
python-dotenv==1.0.1
